// src/app/interfaces/pupil.interface.ts
export interface PupilData {
    id: string;
    first_name: string;
    last_name: string;
    avatar_name: string;
    avatar_url: any;
    father_mobile: string;
    mother_mobile: string;
    date_of_birth: string;
    home_address: string;
    allergies: string;
    blood_group: string;
    group_class: string;
    meal: string;
    tutor: string;
    sex: string;
    status: number;
    created_at: number;
    updated_at: number;
}